// src/api-client/models/promotion-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PromotionRequest
 */
export interface PromotionRequest {
    /**
     * 
     * @type {string}
     * @memberof PromotionRequest
     */
    'code': string;
    /**
     * 
     * @type {string}
     * @memberof PromotionRequest
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof PromotionRequest
     */
    'discountPercent': number;
    /**
     * 
     * @type {string}
     * @memberof PromotionRequest
     */
    'startDate': string;
    /**
     * 
     * @type {string}
     * @memberof PromotionRequest
     */
    'endDate': string;
}